Commonize the bootctrl HAL
Why "commonize" ? It's because the current bootctrl HAL is focusing on
supporting Grub. From the file name to the place where it read the config
to alter it all contain "grub" on it.

Renaming files & functions to not mentioning Grub again. Also allow specifying
which file to read using a property.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>